Add burger menu to layout.tsx
