+++ apps/customer-website/src/app/page.tsx
@@ -1,60 +1,39 @@
 "use client";
 
-import React, { useState } from "react";
-import Header from "../components/figma/Menu/Header";
-import BurgerMenu from "../components/figma/Menu/BurgerMenu";
+import React from "react";
 import { FeaturedSection } from '@/components/figma/FeaturedSection';
 import RightPicsSmall from "@/components/figma/FeaturedSection/RightPicsSmall";
 import LeftPicsSmall from "@/components/figma/FeaturedSection/LeftPicsSmall";
 
 
 export default function Home() {
-  const [isBurgerMenuOpen, setIsBurgerMenuOpen] = useState(false);
 
-  const handleBurgerMenuToggle = (isOpen: boolean) => {
-    setIsBurgerMenuOpen(isOpen);
-    console.log('Burger menu state:', isOpen); // For debugging
-  };
   return (
-    <div className="w-full flex flex-col bg-[#05050a] relative overflow-x-hidden min-h-screen">
+    <>
+      {/* Main content */}
+      <main className="flex flex-col py-8">
 
-      <Header onBurgerMenuToggle={handleBurgerMenuToggle} />
+        {/* Featured Items Section */}
+        <section className="pt-16">
+          <FeaturedSection />
+        </section>
 
-      {/* Mobile Menu Overlay - shows when burger menu is open */}
-      {isBurgerMenuOpen ? (
-        <BurgerMenu />
-      ) : (
-        <>
-          {/* Main content */}
-          <main className="flex flex-col py-8">
+        <div className="content-stretch flex flex-row items-start justify-between relative size-full sm:hidden">
+          <div className="grid">
+            <LeftPicsSmall />
+          </div>
+          <div className="grid">
+            <RightPicsSmall />
+          </div>
+        </div>
 
-            {/* Featured Items Section */}
-            <section className="pt-16">
-              <FeaturedSection />
-            </section>
+        {/* Reviews Section */}
+        {/* <ReviewsSection /> */}
+      </main>
 
-            <div className="content-stretch flex flex-row items-start justify-between relative size-full sm:hidden">
+      {/* Responzívny Footer */}
+      {/* <ResponsiveFooter /> */}
+    </>
 
-              {/* <div className="grid grid-cols-[max-content] grid-rows-[max-content] inline-grid place-items-start relative shrink-0 md:hidden"> */}
-                <div className="grid">
-                  <LeftPicsSmall />
-                </div>
-                <div className="grid">
-                  <RightPicsSmall />
-                </div>
-              {/* </div> */}
-
-            </div>
-
-            {/* Reviews Section */}
-            {/* <ReviewsSection /> */}
-          </main>
-
-          {/* Responzívny Footer */}
-          {/* <ResponsiveFooter /> */}
-        </>
-      )}
-
-    </div>
   );
 }
