+++ apps/customer-website/src/components/figma/Menu/BurgerMenu.tsx
@@ -8,19 +8,13 @@ import ButtonSecondary from '../ButtonSecondary';
 interface BurgerMenuProps {
   /** Whether the burger menu is open */
   isOpen?: boolean;
-  /** Callback function when menu should be closed */
-  onClose?: () => void;
 }
 
-export default function BurgerMenu({ isOpen = true, onClose }: BurgerMenuProps) {
-//   const handleLinkClick = () => {
-//     onClose?.();
-//   };
+export default function BurgerMenu({ isOpen = true }: BurgerMenuProps) {
 
   if (!isOpen) return null;
 
   return (
-    // <div className="lg:hidden fixed inset-0 z-50 bg-[#05050a]" data-node-id="4224:11821">
       <div className="bg-[#05050a] box-border content-stretch flex flex-col gap-[16px] items-center justify-center pb-[240px] pt-[160px] px-0 relative rounded-bl-[40px] rounded-br-[40px] size-full">
         <div className="content-stretch flex flex-col gap-[8px] items-center relative shrink-0" data-node-id="10945:24580">
           
@@ -29,7 +23,6 @@ export default function BurgerMenu({ isOpen = true, onClose }: BurgerMenuProps) 
             href="/vozidla" 
             className="box-border content-stretch flex gap-[16px] h-[40px] items-center justify-center px-[24px] py-[16px] relative shrink-0 hover:bg-[#1a1a1a] transition-colors rounded-[4px]" 
             data-node-id="10945:24581"
-            // onClick={handleLinkClick}
           >
             <div className="flex flex-col font-poppins font-medium justify-center leading-[0] not-italic relative shrink-0 text-[#a0a0a5] text-[16px] text-center text-nowrap hover:text-[#d7ff14] transition-colors" data-node-id="10945:24582">
               <p className="leading-[24px] whitespace-pre">Ponuka vozidiel</p>
@@ -40,7 +33,6 @@ export default function BurgerMenu({ isOpen = true, onClose }: BurgerMenuProps) 
             href="/store" 
             className="box-border content-stretch flex gap-[16px] h-[40px] items-center justify-center px-[24px] py-[16px] relative shrink-0 hover:bg-[#1a1a1a] transition-colors rounded-[4px]" 
             data-node-id="10945:24583"
-            // onClick={handleLinkClick}
           >
             <div className="flex flex-col font-poppins font-medium justify-center leading-[0] not-italic relative shrink-0 text-[#a0a0a5] text-[16px] text-center text-nowrap hover:text-[#d7ff14] transition-colors" data-node-id="10945:24584">
               <p className="leading-[24px] whitespace-pre">Store</p>
@@ -51,7 +43,6 @@ export default function BurgerMenu({ isOpen = true, onClose }: BurgerMenuProps) 
             href="/sluzby" 
             className="box-border content-stretch flex gap-[16px] h-[40px] items-center justify-center px-[24px] py-[16px] relative shrink-0 hover:bg-[#1a1a1a] transition-colors rounded-[4px]" 
             data-node-id="10945:24585"
-            // onClick={handleLinkClick}
           >
             <div className="flex flex-col font-poppins font-medium justify-center leading-[0] not-italic relative shrink-0 text-[#a0a0a5] text-[16px] text-center text-nowrap hover:text-[#d7ff14] transition-colors" data-node-id="10945:24586">
               <p className="leading-[24px] whitespace-pre">Služby</p>
@@ -62,7 +53,6 @@ export default function BurgerMenu({ isOpen = true, onClose }: BurgerMenuProps) 
             href="/o-nas" 
             className="box-border content-stretch flex gap-[16px] h-[40px] items-center justify-center px-[24px] py-[16px] relative shrink-0 hover:bg-[#1a1a1a] transition-colors rounded-[4px]" 
             data-node-id="10945:24587"
-            // onClick={handleLinkClick}
           >
             <div className="flex flex-col font-poppins font-medium justify-center leading-[0] not-italic relative shrink-0 text-[#a0a0a5] text-[16px] text-center text-nowrap hover:text-[#d7ff14] transition-colors" data-node-id="10945:24588">
               <p className="leading-[24px] whitespace-pre">O nás</p>
@@ -73,7 +63,6 @@ export default function BurgerMenu({ isOpen = true, onClose }: BurgerMenuProps) 
             href="/kontakt" 
             className="box-border content-stretch flex gap-[16px] h-[40px] items-center justify-center px-[24px] py-[16px] relative shrink-0 hover:bg-[#1a1a1a] transition-colors rounded-[4px]" 
             data-node-id="10945:24589"
-            // onClick={handleLinkClick}
           >
             <div className="flex flex-col font-poppins font-medium justify-center leading-[0] not-italic relative shrink-0 text-[#a0a0a5] text-[16px] text-center text-nowrap hover:text-[#d7ff14] transition-colors" data-node-id="10945:24590">
               <p className="leading-[24px] whitespace-pre">Kontakt</p>
@@ -99,11 +88,9 @@ export default function BurgerMenu({ isOpen = true, onClose }: BurgerMenuProps) 
               iconAlt="Prihlásiť sa"
               text="Prihlásiť sa"
               className="bg-[#141900] hover:bg-[#1f2600]"
-            //   onClick={handleLinkClick}
             />
           </div>
         </div>
       </div>
-     // </div>
   );
 }
